moved & corrected bf_operator.h
